@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace helloworld_csharp
 {
@@ -15,17 +16,22 @@ namespace helloworld_csharp
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
-            app.Run(async (context) =>
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
             {
-                var target = Environment.GetEnvironmentVariable("TARGET") ?? "World";
-                await context.Response.WriteAsync($"Hello {target}!\n");
+                endpoints.MapGet("/", async context =>
+                {
+                    var target = Environment.GetEnvironmentVariable("TARGET") ?? "World";
+                    await context.Response.WriteAsync($"Hello {target}!\n");
+                });
             });
         }
     }
