@@ -28,9 +28,9 @@ import (
 )
 
 var (
-	yamlFile     = flag.String("yaml", "eventing/sources/sources.yaml", "The YAML file to parse to generate the mark down.")
-	templateFile = flag.String("template", "eventing/sources/generator/ReadmeTemplate.gomd", "The template file to fill in.")
-	mdFile       = flag.String("md", "eventing/sources/README.md", "The mark down file to write to. Any existing file will be overwritten.")
+	yamlFile     = flag.String("yaml", "docs/eventing/sources/sources.yaml", "The YAML file to parse to generate the mark down.")
+	templateFile = flag.String("template", "docs/eventing/sources/generator/ReadmeTemplate.gomd", "The template file to fill in.")
+	mdFile       = flag.String("md", "docs/eventing/sources/README.md", "The mark down file to write to. Any existing file will be overwritten.")
 )
 
 func main() {
@@ -57,6 +57,7 @@ func parseYaml() *yamlSources {
 	sortAlphabetically(sources.MetaSources)
 	sortAlphabetically(sources.Sources)
 	sortAlphabetically(sources.Containers)
+	sortAlphabetically(sources.SinkBindings)
 
 	return sources
 }
@@ -69,9 +70,10 @@ func sortAlphabetically(slice []source) {
 }
 
 type yamlSources struct {
-	MetaSources []source `yaml:"metaSources"`
-	Sources     []source `yaml:"sources"`
-	Containers  []source `yaml:"containers"`
+	MetaSources  []source `yaml:"metaSources"`
+	Sources      []source `yaml:"sources"`
+	Containers   []source `yaml:"containers"`
+	SinkBindings []source `yaml:"sinkBindings"`
 }
 
 type source struct {
